refac: markets helper
